feat: return only `FieldInfo`s from `fields_from`
The original `fields_from` function was renamed to `fields_from_deprecated` and was marked as deprecated. It is still available without any changes from the user, though. It will be removed in a future release.

The new `fields_from` function only returns a `dict` containing all the `FieldInfo`s parsed from the SQLAlchemy model. This dictionary can then be used to declare new Pydantic models as usual.